--- nym-proxy/src/lib.rs
@@ -1,9 +1,0 @@
-//! Nym wrapper for lightwallet service RPC services.
-//! TODO: Cratify when development starts proper.
-
-#![warn(missing_docs)]
-#![forbid(unsafe_code)]
-
-pub mod proxy;
-pub mod server;
-pub mod utils;
