+++ zingo-nymd/src/lib.rs
@@ -1,0 +1,7 @@
+//! Nym-Proxy gRPC server
+
+#![warn(missing_docs)]
+#![forbid(unsafe_code)]
+
+pub mod proxy;
+pub mod server;
