+++ zingo-nymd/zcash.conf
