+++ zingo-rpc/src/nym.rs
@@ -1,0 +1,3 @@
+//! Backend Nym functionality.
+
+pub mod utils;
